add status checking after calling solc compiler; discard output to /dev/null instead of printing in console

+++ resources/tests/YulTests/EmptyContract.sh
@@ -6,6 +6,12 @@ sbt "runMain edu.cmu.cs.obsidian.Main --yul resources/tests/YulTests/EmptyContra
 
 cd EmptyContract
 
-solc --strict-assembly EmptyContract.yul
+solc --strict-assembly EmptyContract.yul > /dev/null
+
+if [ $? -ne 0 ]; then
+  echo "EmptyContract test failed: solc cannot compile yul code"
+  exit 1
+fi
+
 
 rm -rf EmptyContract
